Rounding errors solved by using BigDecimal

+++ src/main/java/com/N26/robotfactory/adapter/EmptyRobot.java
@@ -3,13 +3,14 @@ package com.N26.robotfactory.adapter;
 import com.N26.robotfactory.domain.model.ComponentInventory;
 import com.N26.robotfactory.gateway.IRobot;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public class EmptyRobot implements IRobot {
 
 
     @Override
-    public Double findPrice(List<ComponentInventory> componentInventory,String productCode) {
+    public BigDecimal findPrice(List<ComponentInventory> componentInventory, String productCode) {
 
         return null;
     }
