+++ src/main/java/com/N26/robotfactory/adapter/RobotArm.java
@@ -4,6 +4,7 @@ import com.N26.robotfactory.data.StockRepository;
 import com.N26.robotfactory.domain.model.ComponentInventory;
 import com.N26.robotfactory.gateway.IRobot;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public class RobotArm implements IRobot {
@@ -11,7 +12,7 @@ public class RobotArm implements IRobot {
     StockRepository stockRepository = new StockRepository();
 
     @Override
-    public Double findPrice(List<ComponentInventory> componentInventory, String productCode) {
+    public BigDecimal findPrice(List<ComponentInventory> componentInventory, String productCode) {
 
         return stockRepository.findRobotPartStockByCode(componentInventory, productCode).getPrice();
     }
